create git attributes file
Means we can track the large files in the unity project without them having to be in the repository. Not sure if this fixes the issue of using git add . or not. Will have to see if this works as I think it does otherwise I will remove this file from the repository.